requirements.txt
numpy>=1.21.5,<1.27
pandas>=1.3.5,<1.6

# Kafka integration
kafka-python>=2.0.2
confluent-kafka>=1.8.2

# Monitoring and metrics
prometheus-client>=0.13.1
psutil>=5.9.0

# Utilities
pyyaml>=6.0
